Move types module into protocol
