+++ src/protocol/types/hash.rs
