+++ src/server/mod.rs
@@ -21,14 +21,14 @@ use std::sync::mpsc;
 use std::thread;
 use std::collections::HashMap;
 use std::hash::BuildHasherDefault;
-use crate::types::hash::FNVHash;
+use crate::protocol::types::hash::FNVHash;
 use crate::resources;
 use crate::render;
 use crate::settings::Stevenkey;
 use crate::ecs;
 use crate::entity;
 use cgmath::prelude::*;
-use crate::types::Gamemode;
+use crate::protocol::types::Gamemode;
 use crate::shared::{Axis, Position};
 use crate::protocol::format;
 use rsa_public_encrypt_pkcs1;
